+++ cache-config/t3c-apply/t3c-apply.go
@@ -165,6 +165,10 @@ func Main() int {
 				log.Errorf("couldn't remove git lock file: %v", err.Error())
 			}
 		}
+		log.Infoln("Checking git for safe directory config")
+		if err := util.GetGitConfigSafeDir(cfg); err != nil {
+			log.Warnln("error checking git for safe directory config: " + err.Error())
+		}
 		// commit anything someone else changed when we weren't looking,
 		// with a keyword indicating it wasn't our change
 		if err := util.MakeGitCommitAll(cfg, util.GitChangeNotSelf, true); err != nil {
@@ -256,6 +260,7 @@ func Main() int {
 	if cfg.Files != t3cutil.ApplyFilesFlagAll {
 		// make sure we got the data necessary to check packages
 		log.Infoln("======== Didn't get all files, no package processing needed or possible ========")
+		metaData.InstalledPackages = oldMetaData.InstalledPackages
 	} else {
 		log.Infoln("======== Start processing packages  ========")
 		err = trops.ProcessPackages()
@@ -263,6 +268,7 @@ func Main() int {
 			log.Errorf("Error processing packages: %s\n", err)
 			return GitCommitAndExit(ExitCodePackagingError, FailureExitMsg, cfg, metaData, oldMetaData)
 		}
+		metaData.InstalledPackages = t3cutil.PackagesToMetaData(trops.Pkgs)
 
 		// check and make sure packages are enabled for startup
 		err = trops.CheckSystemServices()
@@ -365,6 +371,9 @@ func GitCommitAndExit(exitCode int, exitMsg string, cfg config.Cfg, metaData *t3
 	// so add the old files to the new metadata.
 	// This is especially important for reval runs, which don't add all files.
 	metaData.OwnedFilePaths = t3cutil.CombineOwnedFilePaths(metaData, oldMetaData)
+	if len(metaData.InstalledPackages) == 0 {
+		metaData.InstalledPackages = oldMetaData.InstalledPackages
+	}
 	WriteMetaData(cfg, metaData)
 	success := exitCode == ExitCodeSuccess
 	if cfg.UseGit == config.UseGitYes || cfg.UseGit == config.UseGitAuto {