+++ traffic_monitor/cache/cache_test.go
@@ -21,14 +21,15 @@ package cache
 
 import (
 	"bytes"
-	"fmt"
+	"io/ioutil"
+	"net/http"
+	"testing"
+
+	"github.com/apache/trafficcontrol/lib/go-rfc"
 	"github.com/apache/trafficcontrol/lib/go-tc"
 	"github.com/apache/trafficcontrol/lib/go-util"
 	"github.com/apache/trafficcontrol/traffic_monitor/poller"
 	"github.com/apache/trafficcontrol/traffic_monitor/todata"
-	"io/ioutil"
-	"net/http"
-	"testing"
 )
 
 func TestHandlerPrecompute(t *testing.T) {
@@ -108,7 +109,7 @@ func TestParseAndDecode(t *testing.T) {
 
 	pl := &poller.HTTPPollCtx{HTTPHeader: http.Header{}}
 	ctx := interface{}(pl)
-	ctx.(*poller.HTTPPollCtx).HTTPHeader.Set("Content-Type", "text/json")
+	ctx.(*poller.HTTPPollCtx).HTTPHeader.Set(rfc.ContentType, rfc.ApplicationJSON)
 
 	decoder, err := GetDecoder("stats_over_http")
 	if err != nil {
@@ -125,8 +126,7 @@ func TestParseAndDecode(t *testing.T) {
 	}
 
 	if val, ok := miscStats["plugin.system_stats.timestamp_ms"]; ok {
-		valString := fmt.Sprintf("%s", val)
-		if valString != "1684784877939" {
+		if val.(string) != "1684784877939" {
 			t.Errorf("unable to read `plugin.system_stats.timestamp_ms`")
 		}
 	} else {