+++ traffic_ops/traffic_ops_golang/dbhelpers/db_helpers_test.go
@@ -562,3 +562,54 @@ func TestASNExists(t *testing.T) {
 		})
 	}
 }
+
+func TestDivisionExists(t *testing.T) {
+	var testCases = []struct {
+		description   string
+		id            string
+		expectedError error
+		exists        bool
+	}{
+		{
+			description:   "Success: Get valid Division",
+			id:            "1",
+			expectedError: nil,
+			exists:        true,
+		},
+		{
+			description:   "Failure: Division not in DB",
+			id:            "10",
+			expectedError: sql.ErrNoRows,
+			exists:        false,
+		},
+	}
+	for _, testCase := range testCases {
+		t.Run(testCase.description, func(t *testing.T) {
+			mockDB, mock, err := sqlmock.New()
+			if err != nil {
+				t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
+			}
+			defer mockDB.Close()
+
+			db := sqlx.NewDb(mockDB, "sqlmock")
+			defer db.Close()
+
+			mock.ExpectBegin()
+			rows := sqlmock.NewRows([]string{"count"})
+			if testCase.exists {
+				rows = rows.AddRow(1)
+			}
+			mock.ExpectQuery("SELECT").WillReturnRows(rows)
+			mock.ExpectCommit()
+
+			divisionExists, err := DivisionExists(db.MustBegin().Tx, testCase.id)
+			if testCase.exists != divisionExists {
+				t.Errorf("Expected return exists: %t, actual %t", testCase.exists, divisionExists)
+			}
+
+			if !errors.Is(err, testCase.expectedError) {
+				t.Errorf("DivisionExists Error. expected: %s, actual: %s", testCase.expectedError, err)
+			}
+		})
+	}
+}